backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
boto3==1.34.34
pydantic==1.10.13
SQLAlchemy==2.0.29
alembic==1.13.1
psycopg[binary]==3.1.18
redis==5.0.1
celery==5.3.6
python-dotenv==1.0.1
packaging==24.0
llama-cpp-python==0.2.56
pytest==8.1.1
pytest-asyncio==0.23.5
httpx==0.27.0
ruff==0.3.2
mypy==1.8.0
bandit==1.7.7
